CAMEL-22894: Fix pre-existing bugs surfaced by SimpleExpressionBuilder split
- shuffleExpression.toString() returned "reverse(...)" due to a copy-paste
  error; now returns "shuffle(...)"
- mapExpression.toString() printed compiled Expression[] object refs instead
  of the source String[] pairs; now prints the source pairs
- collateExpression(String,String) overwrote the 'exp' field inside
  evaluate(), wrapping an already-grouped iterator on every call and causing
  a data race under concurrent route execution; fixed by using a local
  variable for the grouped expression
- throwExceptionExpression's catch(Exception) block caught the RuntimeException
  it had just thrown and double-wrapped it, breaking typed onException handlers;
  restructured so RuntimeExceptions propagate unwrapped
- skipExpression(String,String) unboxed a potentially-null Integer without a
  null check; now throws IllegalArgumentException with a clear message
- padExpression returned NPE when the string or width expression was null;
  now returns null for a null string and throws IllegalArgumentException for
  a null width
- floor() and ceil() cast Math.floor/ceil to int, silently overflowing for
  values outside Integer range; now returns long
- range(max) in the csimple code path used min=0 while the simple interpreter
  used min=1; unified to min=1 to match documented behaviour

Tests added to document each pre-existing bug and verify the fix.


rh-pre-commit.version: 2.4.0
rh-pre-commit.check-secrets: ENABLED

+++ core/camel-core-languages/src/main/java/org/apache/camel/language/simple/MathExpressionBuilder.java
@@ -99,7 +99,7 @@ public final class MathExpressionBuilder {
                 }
                 if (value != null) {
                     double d = Math.floor(value);
-                    return (int) d;
+                    return (long) d;
                 }
                 return value;
             }
@@ -141,7 +141,7 @@ public final class MathExpressionBuilder {
                 }
                 if (value != null) {
                     double d = Math.ceil(value);
-                    return (int) d;
+                    return (long) d;
                 }
                 return value;
             }
